@@ -1,53 +1,37 @@
 {
-  "name": "chat_log_web",
-  "version": "0.1.0",
+  "name": "test",
+  "version": "0.0.0",
   "private": true,
+  "type": "module",
   "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "export NODE_OPTIONS=--openssl-legacy-provider && vue-cli-service build",
-    "build-qiniu": "vue-cli-service build && node ./build_qiniu.js",
-    "build-server": "npm run build && node ./build_server.js",
-    "lint": "vue-cli-service lint"
+    "dev": "vite",
+    "build": "run-p type-check \"build-only {@}\" --",
+    "build-server": "vue-cli-service build && node ./build_server.js",
+    "preview": "vite preview",
+    "build-only": "vite build",
+    "type-check": "vue-tsc --build --force"
   },
   "dependencies": {
     "@vicons/ionicons5": "^0.12.0",
-    "axios": "^0.26.0",
-    "compressing": "^1.6.2",
-    "core-js": "^3.6.5",
-    "qiniu": "^7.6.1",
-    "qs": "^6.11.0",
-    "ssh2": "^1.11.0",
-    "vue": "^3.0.0",
-    "vue-router": "^4.0.14"
+    "@vue/runtime-dom": "^3.4.37",
+    "axios": "^1.7.4",
+    "compressing": "^1.10.1",
+    "md-editor-v3": "^4.18.1",
+    "ssh2": "^1.15.0",
+    "vicons": "^0.0.1",
+    "vue": "^3.4.37",
+    "vue-router": "^4.3.3"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "~4.5.15",
-    "@vue/cli-plugin-eslint": "~4.5.15",
-    "@vue/cli-service": "~4.5.15",
-    "@vue/compiler-sfc": "^3.0.0",
-    "babel-eslint": "^10.1.0",
-    "eslint": "^6.7.2",
-    "eslint-plugin-vue": "^7.0.0",
-    "naive-ui": "^2.33.2",
-    "vfonts": "^0.0.3"
-  },
-  "eslintConfig": {
-    "root": true,
-    "env": {
-      "node": true
-    },
-    "extends": [
-      "plugin:vue/vue3-essential",
-      "eslint:recommended"
-    ],
-    "parserOptions": {
-      "parser": "babel-eslint"
-    },
-    "rules": {}
-  },
-  "browserslist": [
-    "> 1%",
-    "last 2 versions",
-    "not dead"
-  ]
+    "@tsconfig/node20": "^20.1.4",
+    "@types/node": "^20.14.5",
+    "@vitejs/plugin-vue": "^5.0.5",
+    "@vue/tsconfig": "^0.5.1",
+    "naive-ui": "^2.39.0",
+    "npm-run-all2": "^6.2.0",
+    "typescript": "~5.4.0",
+    "vite": "^5.3.1",
+    "vite-plugin-vue-devtools": "^7.3.1",
+    "vue-tsc": "^2.0.21"
+  }
 }
